Fixed regression from last PR.
